% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_EFF_values.R
\name{check_EFF_values}
\alias{check_EFF_values}
\title{check_EFF_values - Checks for unusual/non-standard mesh sizes}
\usage{
check_EFF_values()
}
\value{

}
\description{
Most of the gillnets use common sizes. This function checks for sizes not commonly used.
}
