% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flen_lt_tlen.R
\name{check_flen_lt_tlen}
\alias{check_flen_lt_tlen}
\title{check_flen_lt_tlen - Checks FLEN less than TLEN}
\usage{
check_flen_lt_tlen()
}
\value{
dataframe Table of records that fail the test is returned
}
\description{
Query checks for instances where TLEN is less then FLEN
}
