% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fn2_tables.R
\name{check_fn2_tables}
\alias{check_fn2_tables}
\title{check_fn2_tables}
\usage{
check_fn2_tables()
}
\value{
message Provides usethis messages verifying whether a table exists in memory
}
\description{
This function checks whether all the template tables exist in memory.
}
