% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn121_column_check.R
\name{fn121_column_check}
\alias{fn121_column_check}
\title{Check FN121 Columns}
\usage{
fn121_column_check(FN121)
}
\arguments{
\item{FN121}{FN121 table that adheres to the FN2 data model}
}
\description{
Check FN121 Columns
}
\examples{
\dontrun{
fn121_column_check(FN121)
}

}
