% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn121_column_check.R
\name{fn121_column_check_t5}
\alias{fn121_column_check_t5}
\title{Check FN121 Columns based on GL Template 5}
\usage{
fn121_column_check_t5(FN121)
}
\arguments{
\item{FN121}{FN121 table that adheres to the GL Template 5 data model}
}
\description{
Check FN121 Columns based on GL Template 5
}
\examples{
\dontrun{
fn121_column_check_t5(FN121)
}

}
