% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn121_error_queries.R
\name{fn121_error_queries}
\alias{fn121_error_queries}
\title{FN121 error checks}
\usage{
fn121_error_queries(FN121, FN123)
}
\arguments{
\item{FN121}{FN121 table}

\item{FN123}{FN123 table}
}
\value{
Usethis messages and a list of datafranes that fail tests
}
\description{
FN121 error checks
}
\examples{
#' \dontrun{
fn121_errors <- fn121_error_queries(FN121, FN123)
}
}
