% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn123_column_check.R
\name{fn123_column_check}
\alias{fn123_column_check}
\title{Check FN123 Columns}
\usage{
fn123_column_check(FN123)
}
\arguments{
\item{FN123}{FN123 table that adheres to the FN2 data model}
}
\description{
Check FN123 Columns
}
\examples{
\dontrun{
fn123_column_check(FN123)
}

}
