% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn123_error_queries.R
\name{fn123_error_queries}
\alias{fn123_error_queries}
\title{FN123 Data Check}
\usage{
fn123_error_queries(FN123, FN125)
}
\arguments{
\item{FN123}{FN123 table with PRJ_CD, SAM, EFF, SPC, CATCNT, CATWT, BIOCNT}

\item{FN125}{FN125 table with PRJ_CD, SAM, EFF, SPC, FISH, RWT}
}
\value{
list of named tables with records that should be checked.
}
\description{
FN123 Data Check
}
\examples{
\dontrun{
library(dplyr)
ErrorTables <- fn123_error_queries(FN123, FN125)
lapply(ErrorTables, head)
}
}
