% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template5_tests.R
\name{fn124_minmax_TLEN}
\alias{fn124_minmax_TLEN}
\title{Checks FN124 data against FN012 min/max lengths}
\usage{
fn124_minmax_TLEN(FN124, FN012, makeplot = TRUE)
}
\arguments{
\item{FN124}{FN124 table from Template 5}

\item{FN012}{FN012 table from Template 5}

\item{makeplot}{option to produce plot. Default = TRUE}
}
\value{
FN124 table with length checks from FN012
}
\description{
Checks FN124 data against FN012 min/max lengths
}
