% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn125_fl_tl_lm.R
\name{fn125_fl_tl_lm}
\alias{fn125_fl_tl_lm}
\title{fl_tl_lm - Compares FL:TL using lm relationship}
\usage{
fn125_fl_tl_lm(FN125, makeplot = F, fail_criteria = 0.1)
}
\arguments{
\item{fail_criteria}{}
}
\description{
fl_tl_lm - Compares FL:TL using lm relationship
}
