% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template5_tests.R
\name{fn125_fulton}
\alias{fn125_fulton}
\title{Check FN125 data against FN012 table}
\usage{
fn125_fulton(FN125, FN012, makeplot = TRUE)
}
\arguments{
\item{FN125}{FN125 table from Template 5}

\item{FN012}{FN012 table from Template 5}

\item{makeplot}{option to produce plot. Default = TRUE}
}
\value{
FN125 table with Fulton K, ranges from FN012 and categorization
}
\description{
Check FN125 data against FN012 table
}
\examples{
\dontrun{
FN125 <- fn125_fulton(FN125, FN012, makeplot = T)
FN125 <- fn125_fulton(FN125, FN012, makeplot = F)
}
}
