% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template5_tests.R
\name{fn125_minmax_TLEN}
\alias{fn125_minmax_TLEN}
\title{Check FN125 data against FN012 min/max lengths}
\usage{
fn125_minmax_TLEN(FN125, FN012, makeplot = TRUE)
}
\arguments{
\item{FN125}{FN125 table from Template 5}

\item{FN012}{FN012 table from Template 5}

\item{makeplot}{option to produce plot. Default = TRUE}
}
\value{
FN125 table with length checks from FN012
}
\description{
Check FN125 data against FN012 min/max lengths
}
\examples{
\dontrun{
FN125 <- fn125_minmax_TLEN(FN125, FN012)
FN125 <- fn125_minmax_TLEN(FN125, FN012, makeplot = F)
}
}
