% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn125_tl_rwt_lm.R
\name{fn125_tl_rwt_lm}
\alias{fn125_tl_rwt_lm}
\title{tl_rwt_lm - Compares FL:RWT using nlm relationship}
\usage{
fn125_tl_rwt_lm(FN125, makeplot = F, fail_criteria = 0.3)
}
\arguments{
\item{fail_criteria}{}
}
\value{

}
\description{
tl_rwt_lm - Compares FL:RWT using nlm relationship
}
