% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn125_vonb_check.R
\name{fn125_vonb_check}
\alias{fn125_vonb_check}
\title{vonb_check - Fits VBL growth function to TLEN and AGE}
\usage{
fn125_vonb_check(FN125, makeplot = FALSE, fail_criteria = 0.2)
}
\arguments{
\item{makeplot}{Default = FALSE. If TRUE, requires ggplot2 and will produce plots
of the raw data colour coded to whether it met the pass/fail criteria. VBL fit
also plotted as a line.}

\item{fail_criteria}{Default value = 0.2 The threshold level that if exceeded in
the log ratio of TLEN vs PRED_TLEN will flag the record as a potential error}
}
\value{

}
\description{
vonb_check - Fits VBL growth function to TLEN and AGE
}
\examples{
\dontrun{
library(dplyr)
library(purrr)
FN125 <- FN125 \%>\%
  split(.$SPC) \%>\%
  map(fn125_vonb_check, makeplot = T) \%>\%
  bind_rows()

FN125 <- FN125 \%>\%
  split(.$SPC) \%>\%
  map(fn125_vonb_check, makeplot = T, fail_criteria = 0.1) \%>\%
  bind_rows()
  }

}
