% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn125_vonb_check_flen.R
\name{fn125_vonb_check_flen}
\alias{fn125_vonb_check_flen}
\title{vonb_check_flen - Fits VBL growth function to FLEN and AGE}
\usage{
fn125_vonb_check_flen(FN125, makeplot = FALSE, fail_criteria = 0.2)
}
\arguments{
\item{fail_criteria}{}
}
\value{

}
\description{
vonb_check_flen - Fits VBL growth function to FLEN and AGE
}
