% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template5_tests.R
\name{fn_template5_basic_checks}
\alias{fn_template5_basic_checks}
\title{Basic database structural tests}
\usage{
fn_template5_basic_checks(
  FN012 = FN012,
  FN022 = FN022,
  FN026_Subspace = FN026_Subspace,
  FN028 = FN028,
  FN121 = FN121,
  FN123 = FN123,
  FN124 = FN124,
  FN125 = FN125
)
}
\arguments{
\item{FN125}{}
}
\value{
returns usethis messages
}
\description{
Basic database structural tests
}
\examples{
\dontrun{
fn_template5_basic_checks(FN012, FN022, FN026_Subspace, FN028, FN121, FN123, FN124, FN125)
}

}
